requirements.txt
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
fonttools==4.54.1
h11==0.14.0
idna==3.10
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.3
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pycparser==2.22
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
seaborn==0.13.2
selenium==4.26.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
